\name{twopartqtl-package}
\alias{twopartqtl-package}
\alias{twopartqtl}
\docType{package}
\title{
 QTL Mapping for Point-mass Mixtures}
\description{
Conducts composite interval mapping to identify QTLs in experimental crosses with phenotypes distributed as point-mass mixtures.
Conducts interval mapping with covariates for phenotypes distributed as point-mass mixtures. 
}
\details{
\tabular{ll}{                                                                           
Package: \tab twopartqtl\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-07-22\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Sandra L. Taylor

Maintainer: Sandra L. Taylor <sltaylor@ucdavis.edu>

}
\references{
Taylor, S.L. and K.S. Pollard 20XX. Composite interval mapping to identify 
  quantitative trait loci for point-mass mixture phenotypes.  \emph{Genetics Research}, \bold{XX}, xxx--xxx

}
\keyword{ package }
\seealso{
 \code{\link[qtl]{cim}}, \code{\link[qtl]{scanone}}  
}
